-fixed bugs related to rmtree
